regional-api/lambda/authorizer/requirements.txt
boto3==1.26.0
python-jose==3.3.0
requests==2.28.1